new package structure WIP
