+++ languages/go/host/interfaces.go
@@ -1,0 +1,10 @@
+package host
+
+type Comparer interface {
+	Equal(other Comparer) bool
+	Lt(other Comparer) bool
+}
+
+type Iterator interface {
+	Iter() chan interface{}
+}
