+++ languages/go/oso.go
@@ -1,16 +1,20 @@
 package oso
 
-import "reflect"
+import (
+	"reflect"
+
+	"github.com/osohq/go-oso/host"
+)
 
 type Oso struct {
-	p *Polar
+	p *host.Polar
 }
 
 /*
 Construct a new Oso instance.
 */
 func NewOso() (Oso, error) {
-	if p, e := newPolar(); e != nil {
+	if p, e := host.newPolar(); e != nil {
 		return Oso{}, e
 	} else {
 		return Oso{p: p}, nil
